+++ lib/paper_trail/has_paper_trail.rb
@@ -94,7 +94,7 @@ module PaperTrail
       def serialize_attributes_for_paper_trail(attributes)
         serialized_attributes.each do |key, coder|
           if attributes.key?(key)
-            coder = PaperTrail::Serializers::Yaml unless coder.respond_to?(:dump) # Fall back to YAML if `coder` has no `dump` method
+            coder = PaperTrail::Serializers::YAML unless coder.respond_to?(:dump) # Fall back to YAML if `coder` has no `dump` method
             attributes[key] = coder.dump(attributes[key])
           end
         end
@@ -103,7 +103,7 @@ module PaperTrail
       def unserialize_attributes_for_paper_trail(attributes)
         serialized_attributes.each do |key, coder|
           if attributes.key?(key)
-            coder = PaperTrail::Serializers::Yaml unless coder.respond_to?(:dump)
+            coder = PaperTrail::Serializers::YAML unless coder.respond_to?(:dump)
             attributes[key] = coder.load(attributes[key])
           end
         end
@@ -113,7 +113,7 @@ module PaperTrail
       def serialize_attribute_changes(changes)
         serialized_attributes.each do |key, coder|
           if changes.key?(key)
-            coder = PaperTrail::Serializers::Yaml unless coder.respond_to?(:dump) # Fall back to YAML if `coder` has no `dump` method
+            coder = PaperTrail::Serializers::YAML unless coder.respond_to?(:dump) # Fall back to YAML if `coder` has no `dump` method
             old_value, new_value = changes[key]
             changes[key] = [coder.dump(old_value),
                             coder.dump(new_value)]
@@ -124,7 +124,7 @@ module PaperTrail
       def unserialize_attribute_changes(changes)
         serialized_attributes.each do |key, coder|
           if changes.key?(key)
-            coder = PaperTrail::Serializers::Yaml unless coder.respond_to?(:dump)
+            coder = PaperTrail::Serializers::YAML unless coder.respond_to?(:dump)
             old_value, new_value = changes[key]
             changes[key] = [coder.load(old_value),
                             coder.load(new_value)]