+++ lib/paper_trail/serializers/yaml.rb
@@ -2,15 +2,15 @@ require 'yaml'
 
 module PaperTrail
   module Serializers
-    module Yaml
+    module YAML
       extend self # makes all instance methods become module methods as well
 
       def load(string)
-        YAML.load string
+        ::YAML.load string
       end
 
       def dump(object)
-        YAML.dump object
+        ::YAML.dump object
       end
     end
   end